Controller Runtime POC for ConfigMap Controller
This is a POC that replaces the current ConfigMap controller, which was
built based on the Kubernetes sample controller project, as well as
client/informer/lister code generated using the Kubernetes
code-generator project, with an implementation based on the Kubernetes
controller-runtime project (as used by the Kubebuilder and Operator SDK
projects).  This provides an apples-to-apples comparison of both
approaches for building and implementing controllers, as needed to
determine the best approach for developing additional controllers within
the PostgreSQL Operator.